Associate a UserHandle with each PhoneAccountHandle
This needs to be tracked so that connections get directed to the
correct user. This is not a public API; rather, it will be populated
and read only by the Telecom service.

Expose a new capability, CAPABILITY_MULTI_USER which can be used by
system-privileged apps to indicate that this PhoneAccount may be
used by any user on the device.

Bug: 17421902
Bug: 17421044
Change-Id: I673904a7ae038c88688bba62642e93174e55c308