Bug 18784289: Fix issue with transition animation callback.
This is a regression where we changed the source view for the transition
to dummy view, but since it was not yet attached, it could not get the
handler implicitly to post the onAnimationCompleted callback.  This CL
modifies these ActivityOption transitions to also take an explicit
handler to process the callback on.

Bug: 18784289

Change-Id: I73f745c33b9f8aed91f8d9cd975f37cf7e4128f1