Activity Transition: fix quick back after enter.
Bug 19105460

When an Activity Transition was receiving an exit call
immediately after the enter, the transition for the enter
was still in progress. TransitionManager does not allow
multiple transitions to work at once, so the enter transition
would run, but the exit did not. This CL detects when the
enter transition is still pending and tells the
ActivityTransitionState to delay one frame so that the
enter can finish its required work prior to starting the
exit transition.

Change-Id: I1b40f1e41d61a67da3fd672419ea321e7d0496da