Set installed flag of all users if system app is installed later
If an application(in /data) that other user installed it initially
is appended as system app later, owner user can not use the system app.

The reason this issue occurs is that the owner user's installed flag
is set false when the application was installed by other user
and the installed flag is not updated when the application is appended
as system app later.

So, we fix like as setting the installed flag to true
when the application is appended as system app.

Bug: 28183865
Change-Id: I0b4da5e0bb77fb3baf86d0453e6637a230b104ba